+++ spec/slim_lint/linter/comment_control_statement_spec.rb
@@ -1,0 +1,21 @@
+require 'spec_helper'
+
+describe SlimLint::Linter::CommentControlStatement do
+  include_context 'linter'
+
+  context 'when a control statement contains code' do
+    let(:slim) { '- some_code' }
+
+    it { should_not report_lint }
+  end
+
+  context 'when a control statement contains only a comment' do
+    let(:slim) { <<-SLIM }
+      -# A comment
+      - # Another comment testing leading whitespace
+    SLIM
+
+    it { should report_lint line: 1 }
+    it { should report_lint line: 2 }
+  end
+end